move methods that aren't shared out of AbstractInsertAnalyzedStatement
